repo: consolidate Repo.open logic
