+++ tests/func/test_external_repo.py
@@ -1,7 +1,7 @@
 import os
 from unittest.mock import ANY
 
-from dvc.external_repo import CLONES, external_repo
+from dvc.repo.open_repo import CLONES, external_repo
 from dvc.scm import Git
 from dvc.testing.tmp_dir import make_subrepo
 from dvc.utils import relpath
